separated all the models
