+++ Poloniex/Commands/CancelCommand.cs
@@ -20,9 +20,5 @@ namespace Poloniex
 		}
 	}
 
-	public class OrderCancellationResponse
-	{
-		[JsonProperty("success")]
-		public bool Success { get; set; }
-	}
+	
 }
