+++ Poloniex/Models/OrderCancellationResponse.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Poloniex 
+{
+	public class OrderCancellationResponse
+	{
+		[JsonProperty("success")]
+		public bool Success { get; set; }
+	}
+}
