+++ Poloniex/Models/WithdrawResponse.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Poloniex 
+{
+	public class WithdrawResponse
+	{
+		[JsonProperty("response")]
+		public string Response { get; set; }
+	}
+}
