+++ Poloniex/Queries/GetAvailableAccountBalancesQuery.cs
@@ -48,16 +48,5 @@ namespace Poloniex
 
 			return balances;
 		}
-
-		public class AvailableAccountBalance
-		{
-			public String Symbol { get; set; }
-			
-			public Decimal Exchange { get; set; }
-			
-			public Decimal Margin { get; set; }
-			
-			public Decimal Lending { get; set; }
-		}
 	}
 }
