+++ Poloniex/Queries/GetCompleteBalancesQuery.cs
@@ -24,19 +24,6 @@ namespace Poloniex
 			return response.Values.ToList();
 		}
 
-		public class CompleteBalance
-		{
-			[JsonIgnore]
-			public String Symbol { get; set; }
-
-			[JsonProperty("available")]
-			public Decimal Available { get; set; }
-
-			[JsonProperty("onOrder")]
-			public Decimal OnOrder { get; set; }
-
-			[JsonProperty("btcValue")]
-			public Decimal BtcValue { get; set; }
-		}
+		
 	}
 }
