+++ Poloniex/Queries/GetOpenLoanOffersQuery.cs
@@ -8,7 +8,7 @@ namespace Poloniex
 {
 	public static class GetOpenLoanOffersQuery
 	{
-		public static async Task<IList<OpenLoanOffer>> GetOpenLoanOffersAsync(this PoloniexClient clien)
+		public static async Task<IList<OpenLoanOffer>> GetOpenLoanOffersAsync(this PoloniexClient client)
 		{
 			Dictionary<String, OpenLoanOffer[]> CustomDeserializer(String value)
 			{
@@ -20,7 +20,7 @@ namespace Poloniex
 				return JsonConvert.DeserializeObject<Dictionary<String, OpenLoanOffer[]>>(value);
 			}
 
-			var response = await clien.SendRequestAsync<Dictionary<String, OpenLoanOffer[]>>(new PoloniexRequest
+			var response = await client.SendRequestAsync(new PoloniexRequest
 			{
 				Api = PoloniexApi.Trading,
 				Command = "returnOpenLoanOffers"
@@ -37,28 +37,6 @@ namespace Poloniex
 			}).ToList();
 		}
 
-		public class OpenLoanOffer
-		{
-			[JsonProperty("id")]
-			public Int32 Id { get; set; }
-
-			[JsonIgnore]
-			public String Symbol { get; set; }
-
-			[JsonProperty("date")]
-			public DateTime Date { get; set; }
-
-			[JsonProperty("amount")]
-			public Decimal Amount { get; set; }
-
-			[JsonProperty("duration")]
-			public Int32 Duration { get; set; }
-
-			[JsonProperty("rate")]
-			public Decimal Rate { get; set; }
-
-			[JsonProperty("autoRenew")]
-			public Boolean AutoRenew { get; set; }
-		}
+	
 	}
 }
